Added TempController and moved methods there https://github.com/SoftUni-Internal/suls-issues/issues/3159
