+++ Open Judge System/Web/OJS.Web/Areas/Administration/Controllers/ParticipantsController.cs
@@ -16,10 +16,8 @@
     using Newtonsoft.Json;
 
     using OJS.Common;
-    using OJS.Common.Extensions;
     using OJS.Data;
     using OJS.Data.Models;
-    using OJS.Services.Common.BackgroundJobs.Contracts;
     using OJS.Web.Areas.Administration.Controllers.Common;
     using OJS.Web.Areas.Administration.ViewModels.Participant;
 
@@ -30,12 +28,9 @@
 
     public class ParticipantsController : LecturerBaseGridController
     {
-        private readonly IBackgroundJobService backgroundJobs;
-
-        public ParticipantsController(IOjsData data, IBackgroundJobService backgroundJobs)
+        public ParticipantsController(IOjsData data)
             : base(data)
         {
-            this.backgroundJobs = backgroundJobs;
         }
 
         public override IEnumerable GetData()
@@ -260,38 +255,6 @@
                 .Where(s => s.Processing && !s.Processed && !s.IsDeleted);
 
             allProcessingSubmissions.Update(s => new Submission() { Processing = false });
-
-            return null;
-        }
-
-        // TODO: Remove this method
-        public ActionResult RegisterJobForCleaningSubmissionsForProcessingTable()
-        {
-            string cron = "0 0 * * *";
-            this.backgroundJobs.AddOrUpdateRecurringJob(
-                "CleanSubmissionsForProcessingTable",
-                () => this.CleanSubmissionsForProcessing(),
-                cron);
-
-            return null;
-        }
-
-        // TODO: Remove this method
-        public void CleanSubmissionsForProcessing()
-        {
-            this.Data.Context.SubmissionsForProcessing
-                .Where(s => s.Processed && !s.Processing)
-                .Delete();
-        }
-
-        // TODO: Remove this method after the job is registered
-        public ActionResult RegisterJobForDeletingLeftOverFilesInTempFolder()
-        {
-            var cron = "0 1 * * *";
-            this.backgroundJobs.AddOrUpdateRecurringJob(
-                "DeleteLeftOverFoldersInTempFolder",
-                () => DirectoryHelpers.DeleteExecutionStrategyWorkingDirectories(),
-                cron);
 
             return null;
         }
